% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{BatchGetDocumentsRequest}
\alias{BatchGetDocumentsRequest}
\title{BatchGetDocumentsRequest Object}
\usage{
BatchGetDocumentsRequest(
  newTransaction = NULL,
  documents = NULL,
  readTime = NULL,
  mask = NULL,
  transaction = NULL
)
}
\arguments{
\item{newTransaction}{Starts a new transaction and reads the documents}

\item{documents}{The names of the documents to retrieve}

\item{readTime}{Reads documents as they were at the given time}

\item{mask}{The fields to return}

\item{transaction}{Reads documents in a transaction}
}
\value{
BatchGetDocumentsRequest object
}
\description{
BatchGetDocumentsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Firestore.BatchGetDocuments.
}
\seealso{
Other BatchGetDocumentsRequest functions: 
\code{\link{projects.databases.documents.batchGet}()}
}
\concept{BatchGetDocumentsRequest functions}
