% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{BatchGetDocumentsResponse}
\alias{BatchGetDocumentsResponse}
\title{BatchGetDocumentsResponse Object}
\usage{
BatchGetDocumentsResponse(
  missing = NULL,
  transaction = NULL,
  readTime = NULL,
  found = NULL
)
}
\arguments{
\item{missing}{A document name that was requested but does not exist}

\item{transaction}{The transaction that was started as part of this request}

\item{readTime}{The time at which the document was read}

\item{found}{A document that was requested}
}
\value{
BatchGetDocumentsResponse object
}
\description{
BatchGetDocumentsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The streamed response for Firestore.BatchGetDocuments.
}
\concept{BatchGetDocumentsResponse functions}
