% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{BatchWriteRequest.labels}
\alias{BatchWriteRequest.labels}
\title{BatchWriteRequest.labels Object}
\usage{
BatchWriteRequest.labels()
}
\value{
BatchWriteRequest.labels object
}
\description{
BatchWriteRequest.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels associated with this batch write.
}
\seealso{
Other BatchWriteRequest functions: 
\code{\link{BatchWriteRequest}()},
\code{\link{projects.databases.documents.batchWrite}()}
}
\concept{BatchWriteRequest functions}
