% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{BatchWriteResponse}
\alias{BatchWriteResponse}
\title{BatchWriteResponse Object}
\usage{
BatchWriteResponse(status = NULL, writeResults = NULL)
}
\arguments{
\item{status}{The status of applying the writes}

\item{writeResults}{The result of applying the writes}
}
\value{
BatchWriteResponse object
}
\description{
BatchWriteResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response from Firestore.BatchWrite.
}
\concept{BatchWriteResponse functions}
