% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{CollectionSelector}
\alias{CollectionSelector}
\title{CollectionSelector Object}
\usage{
CollectionSelector(collectionId = NULL, allDescendants = NULL)
}
\arguments{
\item{collectionId}{The collection ID}

\item{allDescendants}{When false, selects only collections that are immediate children of the \code{parent} specified in the containing \code{RunQueryRequest}}
}
\value{
CollectionSelector object
}
\description{
CollectionSelector Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A selection of a collection, such as \verb{messages as m1}.
}
\concept{CollectionSelector functions}
