% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{CommitRequest}
\alias{CommitRequest}
\title{CommitRequest Object}
\usage{
CommitRequest(transaction = NULL, writes = NULL)
}
\arguments{
\item{transaction}{If set, applies all writes in this transaction, and commits it}

\item{writes}{The writes to apply}
}
\value{
CommitRequest object
}
\description{
CommitRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Firestore.Commit.
}
\seealso{
Other CommitRequest functions: 
\code{\link{projects.databases.documents.commit}()}
}
\concept{CommitRequest functions}
