% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{CommitResponse}
\alias{CommitResponse}
\title{CommitResponse Object}
\usage{
CommitResponse(writeResults = NULL, commitTime = NULL)
}
\arguments{
\item{writeResults}{The result of applying the writes}

\item{commitTime}{The time at which the commit occurred}
}
\value{
CommitResponse object
}
\description{
CommitResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Firestore.Commit.
}
\concept{CommitResponse functions}
