% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{DocumentChange}
\alias{DocumentChange}
\title{DocumentChange Object}
\usage{
DocumentChange(targetIds = NULL, removedTargetIds = NULL, document = NULL)
}
\arguments{
\item{targetIds}{A set of target IDs of targets that match this document}

\item{removedTargetIds}{A set of target IDs for targets that no longer match this document}

\item{document}{The new state of the Document}
}
\value{
DocumentChange object
}
\description{
DocumentChange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Document has changed. May be the result of multiple writes, including deletes, that ultimately resulted in a new value for the Document. Multiple DocumentChange messages may be returned for the same logical change, if multiple targets are affected.
}
\concept{DocumentChange functions}
