% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{DocumentDelete}
\alias{DocumentDelete}
\title{DocumentDelete Object}
\usage{
DocumentDelete(readTime = NULL, removedTargetIds = NULL, document = NULL)
}
\arguments{
\item{readTime}{The read timestamp at which the delete was observed}

\item{removedTargetIds}{A set of target IDs for targets that previously matched this entity}

\item{document}{The resource name of the Document that was deleted}
}
\value{
DocumentDelete object
}
\description{
DocumentDelete Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Document has been deleted. May be the result of multiple writes, including updates, the last of which deleted the Document. Multiple DocumentDelete messages may be returned for the same logical delete, if multiple targets are affected.
}
\concept{DocumentDelete functions}
