% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{DocumentMask}
\alias{DocumentMask}
\title{DocumentMask Object}
\usage{
DocumentMask(fieldPaths = NULL)
}
\arguments{
\item{fieldPaths}{The list of field paths in the mask}
}
\value{
DocumentMask object
}
\description{
DocumentMask Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of field paths on a document. Used to restrict a get or update operation on a document to a subset of its fields. This is different from standard field masks, as this is always scoped to a Document, and takes in account the dynamic nature of Value.
}
\concept{DocumentMask functions}
