% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{DocumentRemove}
\alias{DocumentRemove}
\title{DocumentRemove Object}
\usage{
DocumentRemove(removedTargetIds = NULL, document = NULL, readTime = NULL)
}
\arguments{
\item{removedTargetIds}{A set of target IDs for targets that previously matched this document}

\item{document}{The resource name of the Document that has gone out of view}

\item{readTime}{The read timestamp at which the remove was observed}
}
\value{
DocumentRemove object
}
\description{
DocumentRemove Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Document has been removed from the view of the targets. Sent if the document is no longer relevant to a target and is out of view. Can be sent instead of a DocumentDelete or a DocumentChange if the server can not send the new value of the document. Multiple DocumentRemove messages may be returned for the same logical write or delete, if multiple targets are affected.
}
\concept{DocumentRemove functions}
