% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{DocumentsTarget}
\alias{DocumentsTarget}
\title{DocumentsTarget Object}
\usage{
DocumentsTarget(documents = NULL)
}
\arguments{
\item{documents}{The names of the documents to retrieve}
}
\value{
DocumentsTarget object
}
\description{
DocumentsTarget Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A target specified by a set of documents names.
}
\concept{DocumentsTarget functions}
