% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{DocumentTransform}
\alias{DocumentTransform}
\title{DocumentTransform Object}
\usage{
DocumentTransform(fieldTransforms = NULL, document = NULL)
}
\arguments{
\item{fieldTransforms}{The list of transformations to apply to the fields of the document, in order}

\item{document}{The name of the document to transform}
}
\value{
DocumentTransform object
}
\description{
DocumentTransform Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A transformation of a document.
}
\concept{DocumentTransform functions}
