% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{ExistenceFilter}
\alias{ExistenceFilter}
\title{ExistenceFilter Object}
\usage{
ExistenceFilter(targetId = NULL, count = NULL)
}
\arguments{
\item{targetId}{The target ID to which this filter applies}

\item{count}{The total count of documents that match target_id}
}
\value{
ExistenceFilter object
}
\description{
ExistenceFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A digest of all the documents that match a given target.
}
\concept{ExistenceFilter functions}
