% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{FieldFilter}
\alias{FieldFilter}
\title{FieldFilter Object}
\usage{
FieldFilter(value = NULL, field = NULL, op = NULL)
}
\arguments{
\item{value}{The value to compare to}

\item{field}{The field to filter by}

\item{op}{The operator to filter by}
}
\value{
FieldFilter object
}
\description{
FieldFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A filter on a specific field.
}
\concept{FieldFilter functions}
