% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{FieldReference}
\alias{FieldReference}
\title{FieldReference Object}
\usage{
FieldReference(fieldPath = NULL)
}
\arguments{
\item{fieldPath}{No description}
}
\value{
FieldReference object
}
\description{
FieldReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reference to a field, such as \verb{max(messages.time) as max_time}.
}
\concept{FieldReference functions}
