% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{FieldTransform}
\alias{FieldTransform}
\title{FieldTransform Object}
\usage{
FieldTransform(
  setToServerValue = NULL,
  minimum = NULL,
  removeAllFromArray = NULL,
  increment = NULL,
  fieldPath = NULL,
  appendMissingElements = NULL,
  maximum = NULL
)
}
\arguments{
\item{setToServerValue}{Sets the field to the given server value}

\item{minimum}{Sets the field to the minimum of its current value and the given value}

\item{removeAllFromArray}{Remove all of the given elements from the array in the field}

\item{increment}{Adds the given value to the field's current value}

\item{fieldPath}{The path of the field}

\item{appendMissingElements}{Append the given elements in order if they are not already present in the current field value}

\item{maximum}{Sets the field to the maximum of its current value and the given value}
}
\value{
FieldTransform object
}
\description{
FieldTransform Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A transformation of a field of the document.
}
\concept{FieldTransform functions}
