% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{Filter}
\alias{Filter}
\title{Filter Object}
\usage{
Filter(unaryFilter = NULL, compositeFilter = NULL, fieldFilter = NULL)
}
\arguments{
\item{unaryFilter}{A filter that takes exactly one argument}

\item{compositeFilter}{A composite filter}

\item{fieldFilter}{A filter on a document field}
}
\value{
Filter object
}
\description{
Filter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A filter.
}
\concept{Filter functions}
