% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\docType{package}
\name{firestore_googleAuthR}
\alias{firestore_googleAuthR}
\title{Cloud Firestore API
Accesses the NoSQL document database built for automatic scaling, high performance, and ease of application development.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
at 2021-08-19 09:15:03
filename: /home/juha/workspace/autoGoogleAPI//googlefirestorev1.auto/R/firestore_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
\item https://www.googleapis.com/auth/datastore
\item https://www.googleapis.com/auth/cloud-platform
}
}
