% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1ExportDocumentsMetadata}
\alias{GoogleFirestoreAdminV1ExportDocumentsMetadata}
\title{GoogleFirestoreAdminV1ExportDocumentsMetadata Object}
\usage{
GoogleFirestoreAdminV1ExportDocumentsMetadata(
  progressBytes = NULL,
  operationState = NULL,
  startTime = NULL,
  progressDocuments = NULL,
  endTime = NULL,
  collectionIds = NULL,
  outputUriPrefix = NULL
)
}
\arguments{
\item{progressBytes}{The progress, in bytes, of this operation}

\item{operationState}{The state of the export operation}

\item{startTime}{The time this operation started}

\item{progressDocuments}{The progress, in documents, of this operation}

\item{endTime}{The time this operation completed}

\item{collectionIds}{Which collection ids are being exported}

\item{outputUriPrefix}{Where the entities are being exported to}
}
\value{
GoogleFirestoreAdminV1ExportDocumentsMetadata object
}
\description{
GoogleFirestoreAdminV1ExportDocumentsMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for google.longrunning.Operation results from FirestoreAdmin.ExportDocuments.
}
\concept{GoogleFirestoreAdminV1ExportDocumentsMetadata functions}
