% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1ExportDocumentsRequest}
\alias{GoogleFirestoreAdminV1ExportDocumentsRequest}
\title{GoogleFirestoreAdminV1ExportDocumentsRequest Object}
\usage{
GoogleFirestoreAdminV1ExportDocumentsRequest(
  outputUriPrefix = NULL,
  collectionIds = NULL
)
}
\arguments{
\item{outputUriPrefix}{The output URI}

\item{collectionIds}{Which collection ids to export}
}
\value{
GoogleFirestoreAdminV1ExportDocumentsRequest object
}
\description{
GoogleFirestoreAdminV1ExportDocumentsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for FirestoreAdmin.ExportDocuments.
}
\seealso{
Other GoogleFirestoreAdminV1ExportDocumentsRequest functions: 
\code{\link{projects.databases.exportDocuments}()}
}
\concept{GoogleFirestoreAdminV1ExportDocumentsRequest functions}
