% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1ExportDocumentsResponse}
\alias{GoogleFirestoreAdminV1ExportDocumentsResponse}
\title{GoogleFirestoreAdminV1ExportDocumentsResponse Object}
\usage{
GoogleFirestoreAdminV1ExportDocumentsResponse(outputUriPrefix = NULL)
}
\arguments{
\item{outputUriPrefix}{Location of the output files}
}
\value{
GoogleFirestoreAdminV1ExportDocumentsResponse object
}
\description{
GoogleFirestoreAdminV1ExportDocumentsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Returned in the google.longrunning.Operation response field.
}
\concept{GoogleFirestoreAdminV1ExportDocumentsResponse functions}
