% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1Field}
\alias{GoogleFirestoreAdminV1Field}
\title{GoogleFirestoreAdminV1Field Object}
\usage{
GoogleFirestoreAdminV1Field(indexConfig = NULL, name = NULL)
}
\arguments{
\item{indexConfig}{The index configuration for this field}

\item{name}{Required}
}
\value{
GoogleFirestoreAdminV1Field object
}
\description{
GoogleFirestoreAdminV1Field Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a single field in the database. Fields are grouped by their 'Collection Group', which represent all collections in the database with the same id.
}
\seealso{
Other GoogleFirestoreAdminV1Field functions: 
\code{\link{projects.databases.collectionGroups.fields.patch}()}
}
\concept{GoogleFirestoreAdminV1Field functions}
