% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1ImportDocumentsMetadata}
\alias{GoogleFirestoreAdminV1ImportDocumentsMetadata}
\title{GoogleFirestoreAdminV1ImportDocumentsMetadata Object}
\usage{
GoogleFirestoreAdminV1ImportDocumentsMetadata(
  endTime = NULL,
  collectionIds = NULL,
  progressDocuments = NULL,
  progressBytes = NULL,
  operationState = NULL,
  startTime = NULL,
  inputUriPrefix = NULL
)
}
\arguments{
\item{endTime}{The time this operation completed}

\item{collectionIds}{Which collection ids are being imported}

\item{progressDocuments}{The progress, in documents, of this operation}

\item{progressBytes}{The progress, in bytes, of this operation}

\item{operationState}{The state of the import operation}

\item{startTime}{The time this operation started}

\item{inputUriPrefix}{The location of the documents being imported}
}
\value{
GoogleFirestoreAdminV1ImportDocumentsMetadata object
}
\description{
GoogleFirestoreAdminV1ImportDocumentsMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for google.longrunning.Operation results from FirestoreAdmin.ImportDocuments.
}
\concept{GoogleFirestoreAdminV1ImportDocumentsMetadata functions}
