% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1ImportDocumentsRequest}
\alias{GoogleFirestoreAdminV1ImportDocumentsRequest}
\title{GoogleFirestoreAdminV1ImportDocumentsRequest Object}
\usage{
GoogleFirestoreAdminV1ImportDocumentsRequest(
  inputUriPrefix = NULL,
  collectionIds = NULL
)
}
\arguments{
\item{inputUriPrefix}{Location of the exported files}

\item{collectionIds}{Which collection ids to import}
}
\value{
GoogleFirestoreAdminV1ImportDocumentsRequest object
}
\description{
GoogleFirestoreAdminV1ImportDocumentsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for FirestoreAdmin.ImportDocuments.
}
\seealso{
Other GoogleFirestoreAdminV1ImportDocumentsRequest functions: 
\code{\link{projects.databases.importDocuments}()}
}
\concept{GoogleFirestoreAdminV1ImportDocumentsRequest functions}
