% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1Index}
\alias{GoogleFirestoreAdminV1Index}
\title{GoogleFirestoreAdminV1Index Object}
\usage{
GoogleFirestoreAdminV1Index(
  state = NULL,
  queryScope = NULL,
  fields = NULL,
  name = NULL
)
}
\arguments{
\item{state}{Output only}

\item{queryScope}{Indexes with a collection query scope specified allow queries against a collection that is the child of a specific document, specified at query time, and that has the same collection id}

\item{fields}{The fields supported by this index}

\item{name}{Output only}
}
\value{
GoogleFirestoreAdminV1Index object
}
\description{
GoogleFirestoreAdminV1Index Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Firestore indexes enable simple and complex queries against documents in a database.
}
\seealso{
Other GoogleFirestoreAdminV1Index functions: 
\code{\link{projects.databases.collectionGroups.indexes.create}()}
}
\concept{GoogleFirestoreAdminV1Index functions}
