% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1IndexConfig}
\alias{GoogleFirestoreAdminV1IndexConfig}
\title{GoogleFirestoreAdminV1IndexConfig Object}
\usage{
GoogleFirestoreAdminV1IndexConfig(
  reverting = NULL,
  ancestorField = NULL,
  usesAncestorConfig = NULL,
  indexes = NULL
)
}
\arguments{
\item{reverting}{Output only When true, the \code{Field}'s index configuration is in the process of being reverted}

\item{ancestorField}{Output only}

\item{usesAncestorConfig}{Output only}

\item{indexes}{The indexes supported for this field}
}
\value{
GoogleFirestoreAdminV1IndexConfig object
}
\description{
GoogleFirestoreAdminV1IndexConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The index configuration for this field.
}
\concept{GoogleFirestoreAdminV1IndexConfig functions}
