% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1IndexConfigDelta}
\alias{GoogleFirestoreAdminV1IndexConfigDelta}
\title{GoogleFirestoreAdminV1IndexConfigDelta Object}
\usage{
GoogleFirestoreAdminV1IndexConfigDelta(changeType = NULL, index = NULL)
}
\arguments{
\item{changeType}{Specifies how the index is changing}

\item{index}{The index being changed}
}
\value{
GoogleFirestoreAdminV1IndexConfigDelta object
}
\description{
GoogleFirestoreAdminV1IndexConfigDelta Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about an index configuration change.
}
\concept{GoogleFirestoreAdminV1IndexConfigDelta functions}
