% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1IndexField}
\alias{GoogleFirestoreAdminV1IndexField}
\title{GoogleFirestoreAdminV1IndexField Object}
\usage{
GoogleFirestoreAdminV1IndexField(
  arrayConfig = NULL,
  order = NULL,
  fieldPath = NULL
)
}
\arguments{
\item{arrayConfig}{Indicates that this field supports operations on \code{array_value}s}

\item{order}{Indicates that this field supports ordering by the specified order or comparing using =, !=, <, <=, >, >=}

\item{fieldPath}{Can be \strong{name}}
}
\value{
GoogleFirestoreAdminV1IndexField object
}
\description{
GoogleFirestoreAdminV1IndexField Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A field in an index. The field_path describes which field is indexed, the value_mode describes how the field value is indexed.
}
\concept{GoogleFirestoreAdminV1IndexField functions}
