% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1IndexOperationMetadata}
\alias{GoogleFirestoreAdminV1IndexOperationMetadata}
\title{GoogleFirestoreAdminV1IndexOperationMetadata Object}
\usage{
GoogleFirestoreAdminV1IndexOperationMetadata(
  progressBytes = NULL,
  progressDocuments = NULL,
  state = NULL,
  endTime = NULL,
  index = NULL,
  startTime = NULL
)
}
\arguments{
\item{progressBytes}{The progress, in bytes, of this operation}

\item{progressDocuments}{The progress, in documents, of this operation}

\item{state}{The state of the operation}

\item{endTime}{The time this operation completed}

\item{index}{The index resource that this operation is acting on}

\item{startTime}{The time this operation started}
}
\value{
GoogleFirestoreAdminV1IndexOperationMetadata object
}
\description{
GoogleFirestoreAdminV1IndexOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for google.longrunning.Operation results from FirestoreAdmin.CreateIndex.
}
\concept{GoogleFirestoreAdminV1IndexOperationMetadata functions}
