% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1ListFieldsResponse}
\alias{GoogleFirestoreAdminV1ListFieldsResponse}
\title{GoogleFirestoreAdminV1ListFieldsResponse Object}
\usage{
GoogleFirestoreAdminV1ListFieldsResponse(fields = NULL, nextPageToken = NULL)
}
\arguments{
\item{fields}{The requested fields}

\item{nextPageToken}{A page token that may be used to request another page of results}
}
\value{
GoogleFirestoreAdminV1ListFieldsResponse object
}
\description{
GoogleFirestoreAdminV1ListFieldsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for FirestoreAdmin.ListFields.
}
\concept{GoogleFirestoreAdminV1ListFieldsResponse functions}
