% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1ListIndexesResponse}
\alias{GoogleFirestoreAdminV1ListIndexesResponse}
\title{GoogleFirestoreAdminV1ListIndexesResponse Object}
\usage{
GoogleFirestoreAdminV1ListIndexesResponse(indexes = NULL, nextPageToken = NULL)
}
\arguments{
\item{indexes}{The requested indexes}

\item{nextPageToken}{A page token that may be used to request another page of results}
}
\value{
GoogleFirestoreAdminV1ListIndexesResponse object
}
\description{
GoogleFirestoreAdminV1ListIndexesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for FirestoreAdmin.ListIndexes.
}
\concept{GoogleFirestoreAdminV1ListIndexesResponse functions}
