% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1LocationMetadata}
\alias{GoogleFirestoreAdminV1LocationMetadata}
\title{GoogleFirestoreAdminV1LocationMetadata Object}
\usage{
GoogleFirestoreAdminV1LocationMetadata()
}
\value{
GoogleFirestoreAdminV1LocationMetadata object
}
\description{
GoogleFirestoreAdminV1LocationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metadata message for google.cloud.location.Location.metadata.
}
\concept{GoogleFirestoreAdminV1LocationMetadata functions}
