% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleFirestoreAdminV1Progress}
\alias{GoogleFirestoreAdminV1Progress}
\title{GoogleFirestoreAdminV1Progress Object}
\usage{
GoogleFirestoreAdminV1Progress(completedWork = NULL, estimatedWork = NULL)
}
\arguments{
\item{completedWork}{The amount of work completed}

\item{estimatedWork}{The amount of work estimated}
}
\value{
GoogleFirestoreAdminV1Progress object
}
\description{
GoogleFirestoreAdminV1Progress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the progress of the operation. Unit of work is generic and must be interpreted based on where Progress is used.
}
\concept{GoogleFirestoreAdminV1Progress functions}
