% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleLongrunningCancelOperationRequest}
\alias{GoogleLongrunningCancelOperationRequest}
\title{GoogleLongrunningCancelOperationRequest Object}
\usage{
GoogleLongrunningCancelOperationRequest()
}
\value{
GoogleLongrunningCancelOperationRequest object
}
\description{
GoogleLongrunningCancelOperationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for Operations.CancelOperation.
}
\seealso{
Other GoogleLongrunningCancelOperationRequest functions: 
\code{\link{projects.databases.operations.cancel}()}
}
\concept{GoogleLongrunningCancelOperationRequest functions}
