% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{LatLng}
\alias{LatLng}
\title{LatLng Object}
\usage{
LatLng(longitude = NULL, latitude = NULL)
}
\arguments{
\item{longitude}{The longitude in degrees}

\item{latitude}{The latitude in degrees}
}
\value{
LatLng object
}
\description{
LatLng Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An object that represents a latitude/longitude pair. This is expressed as a pair of doubles to represent degrees latitude and degrees longitude. Unless specified otherwise, this object must conform to the WGS84 standard. Values must be within normalized ranges.
}
\concept{LatLng functions}
