% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{ListCollectionIdsRequest}
\alias{ListCollectionIdsRequest}
\title{ListCollectionIdsRequest Object}
\usage{
ListCollectionIdsRequest(pageToken = NULL, pageSize = NULL)
}
\arguments{
\item{pageToken}{A page token}

\item{pageSize}{The maximum number of results to return}
}
\value{
ListCollectionIdsRequest object
}
\description{
ListCollectionIdsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Firestore.ListCollectionIds.
}
\seealso{
Other ListCollectionIdsRequest functions: 
\code{\link{projects.databases.documents.listCollectionIds}()}
}
\concept{ListCollectionIdsRequest functions}
