% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{ListCollectionIdsResponse}
\alias{ListCollectionIdsResponse}
\title{ListCollectionIdsResponse Object}
\usage{
ListCollectionIdsResponse(nextPageToken = NULL, collectionIds = NULL)
}
\arguments{
\item{nextPageToken}{A page token that may be used to continue the list}

\item{collectionIds}{The collection ids}
}
\value{
ListCollectionIdsResponse object
}
\description{
ListCollectionIdsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response from Firestore.ListCollectionIds.
}
\concept{ListCollectionIdsResponse functions}
