% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{ListDocumentsResponse}
\alias{ListDocumentsResponse}
\title{ListDocumentsResponse Object}
\usage{
ListDocumentsResponse(nextPageToken = NULL, documents = NULL)
}
\arguments{
\item{nextPageToken}{The next page token}

\item{documents}{The Documents found}
}
\value{
ListDocumentsResponse object
}
\description{
ListDocumentsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Firestore.ListDocuments.
}
\concept{ListDocumentsResponse functions}
