% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{ListenRequest.labels}
\alias{ListenRequest.labels}
\title{ListenRequest.labels Object}
\usage{
ListenRequest.labels()
}
\value{
ListenRequest.labels object
}
\description{
ListenRequest.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels associated with this target change.
}
\seealso{
Other ListenRequest functions: 
\code{\link{ListenRequest}()},
\code{\link{projects.databases.documents.listen}()}
}
\concept{ListenRequest functions}
