% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{ListenRequest}
\alias{ListenRequest}
\title{ListenRequest Object}
\usage{
ListenRequest(labels = NULL, removeTarget = NULL, addTarget = NULL)
}
\arguments{
\item{labels}{Labels associated with this target change}

\item{removeTarget}{The ID of a target to remove from this stream}

\item{addTarget}{A target to add to this stream}
}
\value{
ListenRequest object
}
\description{
ListenRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request for Firestore.Listen
}
\seealso{
Other ListenRequest functions: 
\code{\link{projects.databases.documents.listen}()}
}
\concept{ListenRequest functions}
