% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{ListenResponse}
\alias{ListenResponse}
\title{ListenResponse Object}
\usage{
ListenResponse(
  documentRemove = NULL,
  documentDelete = NULL,
  targetChange = NULL,
  documentChange = NULL,
  filter = NULL
)
}
\arguments{
\item{documentRemove}{A Document has been removed from a target (because it is no longer relevant to that target)}

\item{documentDelete}{A Document has been deleted}

\item{targetChange}{Targets have changed}

\item{documentChange}{A Document has changed}

\item{filter}{A filter to apply to the set of documents previously returned for the given target}
}
\value{
ListenResponse object
}
\description{
ListenResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Firestore.Listen.
}
\concept{ListenResponse functions}
