% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{MapValue.fields}
\alias{MapValue.fields}
\title{MapValue.fields Object}
\usage{
MapValue.fields()
}
\value{
MapValue.fields object
}
\description{
MapValue.fields Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The map's fields. The map keys represent field names. Field names matching the regular expression \verb{__.*__} are reserved. Reserved field names are forbidden except in certain documented contexts. The map keys, represented as UTF-8, must not exceed 1,500 bytes and cannot be empty.
}
\seealso{
Other MapValue functions: 
\code{\link{MapValue}()}
}
\concept{MapValue functions}
