% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{MapValue}
\alias{MapValue}
\title{MapValue Object}
\usage{
MapValue(MapValue.fields = NULL, fields = NULL)
}
\arguments{
\item{MapValue.fields}{The \link{MapValue.fields} object or list of objects}

\item{fields}{The map's fields}
}
\value{
MapValue object
}
\description{
MapValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map value.
}
\seealso{
Other MapValue functions: 
\code{\link{MapValue.fields}()}
}
\concept{MapValue functions}
