% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{Order}
\alias{Order}
\title{Order Object}
\usage{
Order(field = NULL, direction = NULL)
}
\arguments{
\item{field}{The field to order by}

\item{direction}{The direction to order by}
}
\value{
Order object
}
\description{
Order Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An order on a field.
}
\concept{Order functions}
