% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{PartitionQueryRequest}
\alias{PartitionQueryRequest}
\title{PartitionQueryRequest Object}
\usage{
PartitionQueryRequest(
  partitionCount = NULL,
  pageSize = NULL,
  pageToken = NULL,
  structuredQuery = NULL
)
}
\arguments{
\item{partitionCount}{The desired maximum number of partition points}

\item{pageSize}{The maximum number of partitions to return in this call, subject to \code{partition_count}}

\item{pageToken}{The \code{next_page_token} value returned from a previous call to PartitionQuery that may be used to get an additional set of results}

\item{structuredQuery}{A structured query}
}
\value{
PartitionQueryRequest object
}
\description{
PartitionQueryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Firestore.PartitionQuery.
}
\seealso{
Other PartitionQueryRequest functions: 
\code{\link{projects.databases.documents.partitionQuery}()}
}
\concept{PartitionQueryRequest functions}
