% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{PartitionQueryResponse}
\alias{PartitionQueryResponse}
\title{PartitionQueryResponse Object}
\usage{
PartitionQueryResponse(partitions = NULL, nextPageToken = NULL)
}
\arguments{
\item{partitions}{Partition results}

\item{nextPageToken}{A page token that may be used to request an additional set of results, up to the number specified by \code{partition_count} in the PartitionQuery request}
}
\value{
PartitionQueryResponse object
}
\description{
PartitionQueryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Firestore.PartitionQuery.
}
\concept{PartitionQueryResponse functions}
