% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{Precondition}
\alias{Precondition}
\title{Precondition Object}
\usage{
Precondition(updateTime = NULL, exists = NULL)
}
\arguments{
\item{updateTime}{When set, the target document must exist and have been last updated at that time}

\item{exists}{When set to \code{true}, the target document must exist}
}
\value{
Precondition object
}
\description{
Precondition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A precondition on a document, used for conditional operations.
}
\concept{Precondition functions}
