% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{Projection}
\alias{Projection}
\title{Projection Object}
\usage{
Projection(fields = NULL)
}
\arguments{
\item{fields}{The fields to return}
}
\value{
Projection object
}
\description{
Projection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The projection of document's fields to return.
}
\concept{Projection functions}
