% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.collectionGroups.fields.list}
\alias{projects.databases.collectionGroups.fields.list}
\title{Lists the field configuration and metadata for this database. Currently, FirestoreAdmin.ListFields only supports listing fields that have been explicitly overridden. To issue this query, call FirestoreAdmin.ListFields with the filter set to \code{indexConfig.usesAncestorConfig:false} .}
\usage{
projects.databases.collectionGroups.fields.list(
  parent,
  pageToken = NULL,
  pageSize = NULL,
  filter = NULL
)
}
\arguments{
\item{parent}{Required}

\item{pageToken}{A page token, returned from a previous call to FirestoreAdmin}

\item{pageSize}{The number of results to return}

\item{filter}{The filter to apply to list results}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}
}
