% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.collectionGroups.fields.patch}
\alias{projects.databases.collectionGroups.fields.patch}
\title{Updates a field configuration. Currently, field updates apply only to single field index configuration. However, calls to FirestoreAdmin.UpdateField should provide a field mask to avoid changing any configuration that the caller isn't aware of. The field mask should be specified as: \code{{ paths: 'index_config' }}. This call returns a google.longrunning.Operation which may be used to track the status of the field update. The metadata for the operation will be the type FieldOperationMetadata. To configure the default field settings for the database, use the special \code{Field} with resource name: \verb{projects/\{project_id\}/databases/\{database_id\}/collectionGroups/__default__/fields/*}.}
\usage{
projects.databases.collectionGroups.fields.patch(
  GoogleFirestoreAdminV1Field,
  name,
  updateMask = NULL
)
}
\arguments{
\item{GoogleFirestoreAdminV1Field}{The \link{GoogleFirestoreAdminV1Field} object to pass to this method}

\item{name}{Required}

\item{updateMask}{A mask, relative to the field}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}

Other GoogleFirestoreAdminV1Field functions: 
\code{\link{GoogleFirestoreAdminV1Field}()}
}
\concept{GoogleFirestoreAdminV1Field functions}
