% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.collectionGroups.indexes.create}
\alias{projects.databases.collectionGroups.indexes.create}
\title{Creates a composite index. This returns a google.longrunning.Operation which may be used to track the status of the creation. The metadata for the operation will be the type IndexOperationMetadata.}
\usage{
projects.databases.collectionGroups.indexes.create(
  GoogleFirestoreAdminV1Index,
  parent
)
}
\arguments{
\item{GoogleFirestoreAdminV1Index}{The \link{GoogleFirestoreAdminV1Index} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}

Other GoogleFirestoreAdminV1Index functions: 
\code{\link{GoogleFirestoreAdminV1Index}()}
}
\concept{GoogleFirestoreAdminV1Index functions}
