% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.documents.createDocument}
\alias{projects.databases.documents.createDocument}
\title{Creates a new document.}
\usage{
projects.databases.documents.createDocument(
  Document,
  parent,
  collectionId,
  documentId = NULL,
  mask.fieldPaths = NULL
)
}
\arguments{
\item{Document}{The \link{Document} object to pass to this method}

\item{parent}{Required}

\item{collectionId}{Required}

\item{documentId}{The client-assigned document ID to use for this document}

\item{mask.fieldPaths}{The list of field paths in the mask}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}

Other Document functions: 
\code{\link{Document}()},
\code{\link{projects.databases.documents.patch}()}
}
\concept{Document functions}
