% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.documents.delete}
\alias{projects.databases.documents.delete}
\title{Deletes a document.}
\usage{
projects.databases.documents.delete(
  name,
  currentDocument.updateTime = NULL,
  currentDocument.exists = NULL
)
}
\arguments{
\item{name}{Required}

\item{currentDocument.updateTime}{When set, the target document must exist and have been last updated at that time}

\item{currentDocument.exists}{When set to \code{true}, the target document must exist}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}
}
