% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.documents.get}
\alias{projects.databases.documents.get}
\title{Gets a single document.}
\usage{
projects.databases.documents.get(
  name,
  mask.fieldPaths = NULL,
  readTime = NULL,
  transaction = NULL
)
}
\arguments{
\item{name}{Required}

\item{mask.fieldPaths}{The list of field paths in the mask}

\item{readTime}{Reads the version of the document at the given time}

\item{transaction}{Reads the document in a transaction}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}
}
