% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.documents.list}
\alias{projects.databases.documents.list}
\title{Lists documents.}
\usage{
projects.databases.documents.list(
  parent,
  collectionId,
  pageToken = NULL,
  orderBy = NULL,
  transaction = NULL,
  mask.fieldPaths = NULL,
  pageSize = NULL,
  readTime = NULL,
  showMissing = NULL
)
}
\arguments{
\item{parent}{Required}

\item{collectionId}{Required}

\item{pageToken}{The \code{next_page_token} value returned from a previous List request, if any}

\item{orderBy}{The order to sort results by}

\item{transaction}{Reads documents in a transaction}

\item{mask.fieldPaths}{The list of field paths in the mask}

\item{pageSize}{The maximum number of documents to return}

\item{readTime}{Reads documents as they were at the given time}

\item{showMissing}{If the list should show missing documents}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}
}
