% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.documents.listCollectionIds}
\alias{projects.databases.documents.listCollectionIds}
\title{Lists all the collection IDs underneath a document.}
\usage{
projects.databases.documents.listCollectionIds(
  ListCollectionIdsRequest,
  parent
)
}
\arguments{
\item{ListCollectionIdsRequest}{The \link{ListCollectionIdsRequest} object to pass to this method}

\item{parent}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}

Other ListCollectionIdsRequest functions: 
\code{\link{ListCollectionIdsRequest}()}
}
\concept{ListCollectionIdsRequest functions}
