% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.documents.patch}
\alias{projects.databases.documents.patch}
\title{Updates or inserts a document.}
\usage{
projects.databases.documents.patch(
  Document,
  name,
  mask.fieldPaths = NULL,
  currentDocument.updateTime = NULL,
  currentDocument.exists = NULL,
  updateMask.fieldPaths = NULL
)
}
\arguments{
\item{Document}{The \link{Document} object to pass to this method}

\item{name}{The resource name of the document, for example \code{projects/{project_id}/databases/{database_id}/documents/{document_path}}}

\item{mask.fieldPaths}{The list of field paths in the mask}

\item{currentDocument.updateTime}{When set, the target document must exist and have been last updated at that time}

\item{currentDocument.exists}{When set to \code{true}, the target document must exist}

\item{updateMask.fieldPaths}{The list of field paths in the mask}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}

Other Document functions: 
\code{\link{Document}()},
\code{\link{projects.databases.documents.createDocument}()}
}
\concept{Document functions}
