% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.documents.write}
\alias{projects.databases.documents.write}
\title{Streams batches of document updates and deletes, in order.}
\usage{
projects.databases.documents.write(WriteRequest, database)
}
\arguments{
\item{WriteRequest}{The \link{WriteRequest} object to pass to this method}

\item{database}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}

Other WriteRequest functions: 
\code{\link{WriteRequest.labels}()},
\code{\link{WriteRequest}()}
}
\concept{WriteRequest functions}
