% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.exportDocuments}
\alias{projects.databases.exportDocuments}
\title{Exports a copy of all or a subset of documents from Google Cloud Firestore to another storage system, such as Google Cloud Storage. Recent updates to documents may not be reflected in the export. The export occurs in the background and its progress can be monitored and managed via the Operation resource that is created. The output of an export may only be used once the associated operation is done. If an export operation is cancelled before completion it may leave partial data behind in Google Cloud Storage. For more details on export behavior and output format, refer to: https://cloud.google.com/firestore/docs/manage-data/export-import}
\usage{
projects.databases.exportDocuments(
  GoogleFirestoreAdminV1ExportDocumentsRequest,
  name
)
}
\arguments{
\item{GoogleFirestoreAdminV1ExportDocumentsRequest}{The \link{GoogleFirestoreAdminV1ExportDocumentsRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}

Other GoogleFirestoreAdminV1ExportDocumentsRequest functions: 
\code{\link{GoogleFirestoreAdminV1ExportDocumentsRequest}()}
}
\concept{GoogleFirestoreAdminV1ExportDocumentsRequest functions}
