% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.importDocuments}
\alias{projects.databases.importDocuments}
\title{Imports documents into Google Cloud Firestore. Existing documents with the same name are overwritten. The import occurs in the background and its progress can be monitored and managed via the Operation resource that is created. If an ImportDocuments operation is cancelled, it is possible that a subset of the data has already been imported to Cloud Firestore.}
\usage{
projects.databases.importDocuments(
  GoogleFirestoreAdminV1ImportDocumentsRequest,
  name
)
}
\arguments{
\item{GoogleFirestoreAdminV1ImportDocumentsRequest}{The \link{GoogleFirestoreAdminV1ImportDocumentsRequest} object to pass to this method}

\item{name}{Required}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}

Other GoogleFirestoreAdminV1ImportDocumentsRequest functions: 
\code{\link{GoogleFirestoreAdminV1ImportDocumentsRequest}()}
}
\concept{GoogleFirestoreAdminV1ImportDocumentsRequest functions}
