% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.databases.operations.cancel}
\alias{projects.databases.operations.cancel}
\title{Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns \code{google.rpc.Code.UNIMPLEMENTED}. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to \code{Code.CANCELLED}.}
\usage{
projects.databases.operations.cancel(
  GoogleLongrunningCancelOperationRequest,
  name
)
}
\arguments{
\item{GoogleLongrunningCancelOperationRequest}{The \link{GoogleLongrunningCancelOperationRequest} object to pass to this method}

\item{name}{The name of the operation resource to be cancelled}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}

Other GoogleLongrunningCancelOperationRequest functions: 
\code{\link{GoogleLongrunningCancelOperationRequest}()}
}
\concept{GoogleLongrunningCancelOperationRequest functions}
