% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_functions.R
\name{projects.locations.list}
\alias{projects.locations.list}
\title{Lists information about the supported locations for this service.}
\usage{
projects.locations.list(name, pageSize = NULL, filter = NULL, pageToken = NULL)
}
\arguments{
\item{name}{The resource that owns the locations collection, if applicable}

\item{pageSize}{The maximum number of results to return}

\item{filter}{A filter to narrow down results to a preferred subset}

\item{pageToken}{A page token received from the \code{next_page_token} field in the response}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/datastore
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/datastore)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/firestore}{Google Documentation}
}
