% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{QueryTarget}
\alias{QueryTarget}
\title{QueryTarget Object}
\usage{
QueryTarget(structuredQuery = NULL, parent = NULL)
}
\arguments{
\item{structuredQuery}{A structured query}

\item{parent}{The parent resource name}
}
\value{
QueryTarget object
}
\description{
QueryTarget Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A target specified by a query.
}
\concept{QueryTarget functions}
