% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{ReadOnly}
\alias{ReadOnly}
\title{ReadOnly Object}
\usage{
ReadOnly(readTime = NULL)
}
\arguments{
\item{readTime}{Reads documents at the given time}
}
\value{
ReadOnly object
}
\description{
ReadOnly Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for a transaction that can only be used to read documents.
}
\concept{ReadOnly functions}
