% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{ReadWrite}
\alias{ReadWrite}
\title{ReadWrite Object}
\usage{
ReadWrite(retryTransaction = NULL)
}
\arguments{
\item{retryTransaction}{An optional transaction to retry}
}
\value{
ReadWrite object
}
\description{
ReadWrite Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for a transaction that can be used to read and write documents.
}
\concept{ReadWrite functions}
