% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{RollbackRequest}
\alias{RollbackRequest}
\title{RollbackRequest Object}
\usage{
RollbackRequest(transaction = NULL)
}
\arguments{
\item{transaction}{Required}
}
\value{
RollbackRequest object
}
\description{
RollbackRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Firestore.Rollback.
}
\seealso{
Other RollbackRequest functions: 
\code{\link{projects.databases.documents.rollback}()}
}
\concept{RollbackRequest functions}
