% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{RunQueryRequest}
\alias{RunQueryRequest}
\title{RunQueryRequest Object}
\usage{
RunQueryRequest(
  structuredQuery = NULL,
  transaction = NULL,
  readTime = NULL,
  newTransaction = NULL
)
}
\arguments{
\item{structuredQuery}{A structured query}

\item{transaction}{Reads documents in a transaction}

\item{readTime}{Reads documents as they were at the given time}

\item{newTransaction}{Starts a new transaction and reads the documents}
}
\value{
RunQueryRequest object
}
\description{
RunQueryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Firestore.RunQuery.
}
\seealso{
Other RunQueryRequest functions: 
\code{\link{projects.databases.documents.runQuery}()}
}
\concept{RunQueryRequest functions}
