% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{RunQueryResponse}
\alias{RunQueryResponse}
\title{RunQueryResponse Object}
\usage{
RunQueryResponse(
  transaction = NULL,
  skippedResults = NULL,
  readTime = NULL,
  document = NULL
)
}
\arguments{
\item{transaction}{The transaction that was started as part of this request}

\item{skippedResults}{The number of results that have been skipped due to an offset between the last response and the current response}

\item{readTime}{The time at which the document was read}

\item{document}{A query result}
}
\value{
RunQueryResponse object
}
\description{
RunQueryResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Firestore.RunQuery.
}
\concept{RunQueryResponse functions}
