% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{Target}
\alias{Target}
\title{Target Object}
\usage{
Target(
  resumeToken = NULL,
  targetId = NULL,
  once = NULL,
  readTime = NULL,
  query = NULL,
  documents = NULL
)
}
\arguments{
\item{resumeToken}{A resume token from a prior TargetChange for an identical target}

\item{targetId}{The target ID that identifies the target on the stream}

\item{once}{If the target should be removed once it is current and consistent}

\item{readTime}{Start listening after a specific \code{read_time}}

\item{query}{A target specified by a query}

\item{documents}{A target specified by a set of document names}
}
\value{
Target object
}
\description{
Target Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A specification of a set of documents to listen to.
}
\concept{Target functions}
