% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{TargetChange}
\alias{TargetChange}
\title{TargetChange Object}
\usage{
TargetChange(
  targetChangeType = NULL,
  cause = NULL,
  targetIds = NULL,
  readTime = NULL,
  resumeToken = NULL
)
}
\arguments{
\item{targetChangeType}{The type of change that occurred}

\item{cause}{The error that resulted in this change, if applicable}

\item{targetIds}{The target IDs of targets that have changed}

\item{readTime}{The consistent \code{read_time} for the given \code{target_ids} (omitted when the target_ids are not at a consistent snapshot)}

\item{resumeToken}{A token that can be used to resume the stream for the given \code{target_ids}, or all targets if \code{target_ids} is empty}
}
\value{
TargetChange object
}
\description{
TargetChange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Targets being watched have changed.
}
\concept{TargetChange functions}
