% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{TransactionOptions}
\alias{TransactionOptions}
\title{TransactionOptions Object}
\usage{
TransactionOptions(readOnly = NULL, readWrite = NULL)
}
\arguments{
\item{readOnly}{The transaction can only be used for read operations}

\item{readWrite}{The transaction can be used for both read and write operations}
}
\value{
TransactionOptions object
}
\description{
TransactionOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for creating a new transaction.
}
\concept{TransactionOptions functions}
