% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{UnaryFilter}
\alias{UnaryFilter}
\title{UnaryFilter Object}
\usage{
UnaryFilter(op = NULL, field = NULL)
}
\arguments{
\item{op}{The unary operator to apply}

\item{field}{The field to which to apply the operator}
}
\value{
UnaryFilter object
}
\description{
UnaryFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A filter with a single operand.
}
\concept{UnaryFilter functions}
