% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{Value}
\alias{Value}
\title{Value Object}
\usage{
Value(
  doubleValue = NULL,
  geoPointValue = NULL,
  arrayValue = NULL,
  integerValue = NULL,
  nullValue = NULL,
  bytesValue = NULL,
  booleanValue = NULL,
  timestampValue = NULL,
  mapValue = NULL,
  stringValue = NULL,
  referenceValue = NULL
)
}
\arguments{
\item{doubleValue}{A double value}

\item{geoPointValue}{A geo point value representing a point on the surface of Earth}

\item{arrayValue}{An array value}

\item{integerValue}{An integer value}

\item{nullValue}{A null value}

\item{bytesValue}{A bytes value}

\item{booleanValue}{A boolean value}

\item{timestampValue}{A timestamp value}

\item{mapValue}{A map value}

\item{stringValue}{A string value}

\item{referenceValue}{A reference to a document}
}
\value{
Value object
}
\description{
Value Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message that can hold any of the supported value types.
}
\concept{Value functions}
