% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{Write}
\alias{Write}
\title{Cloud Firestore API Objects
Accesses the NoSQL document database built for automatic scaling, high performance, and ease of application development.}
\usage{
Write(
  transform = NULL,
  updateTransforms = NULL,
  delete = NULL,
  currentDocument = NULL,
  updateMask = NULL,
  update = NULL
)
}
\arguments{
\item{transform}{Applies a transformation to a document}

\item{updateTransforms}{The transforms to perform after update}

\item{delete}{A document name to delete}

\item{currentDocument}{An optional precondition on the document}

\item{updateMask}{The fields to update in this write}

\item{update}{A document to write}
}
\value{
Write object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
at 2021-08-19 09:15:03
filename: /home/juha/workspace/autoGoogleAPI//googlefirestorev1.auto/R/firestore_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Write Object

Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A write on a document.
}
\concept{Write functions}
