% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{WriteRequest}
\alias{WriteRequest}
\title{WriteRequest Object}
\usage{
WriteRequest(writes = NULL, streamToken = NULL, labels = NULL, streamId = NULL)
}
\arguments{
\item{writes}{The writes to apply}

\item{streamToken}{A stream token that was previously sent by the server}

\item{labels}{Labels associated with this write request}

\item{streamId}{The ID of the write stream to resume}
}
\value{
WriteRequest object
}
\description{
WriteRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Firestore.Write. The first request creates a stream, or resumes an existing one from a token. When creating a new stream, the server replies with a response containing only an ID and a token, to use in the next request. When resuming a stream, the server first streams any responses later than the given token, then a response containing only an up-to-date token, to use in the next request.
}
\seealso{
Other WriteRequest functions: 
\code{\link{projects.databases.documents.write}()}
}
\concept{WriteRequest functions}
