% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{WriteResponse}
\alias{WriteResponse}
\title{WriteResponse Object}
\usage{
WriteResponse(
  streamId = NULL,
  streamToken = NULL,
  commitTime = NULL,
  writeResults = NULL
)
}
\arguments{
\item{streamId}{The ID of the stream}

\item{streamToken}{A token that represents the position of this response in the stream}

\item{commitTime}{The time at which the commit occurred}

\item{writeResults}{The result of applying the writes}
}
\value{
WriteResponse object
}
\description{
WriteResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for Firestore.Write.
}
\concept{WriteResponse functions}
