% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{WriteResult}
\alias{WriteResult}
\title{WriteResult Object}
\usage{
WriteResult(updateTime = NULL, transformResults = NULL)
}
\arguments{
\item{updateTime}{The last update time of the document after applying the write}

\item{transformResults}{The results of applying each DocumentTransform}
}
\value{
WriteResult object
}
\description{
WriteResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of applying a write.
}
\concept{WriteResult functions}
