% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{BatchWriteRequest}
\alias{BatchWriteRequest}
\title{BatchWriteRequest Object}
\usage{
BatchWriteRequest(labels = NULL, writes = NULL)
}
\arguments{
\item{labels}{Labels associated with this batch write}

\item{writes}{The writes to apply}
}
\value{
BatchWriteRequest object
}
\description{
BatchWriteRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for Firestore.BatchWrite.
}
\seealso{
Other BatchWriteRequest functions: 
\code{\link{projects.databases.documents.batchWrite}()}
}
\concept{BatchWriteRequest functions}
