% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{Document}
\alias{Document}
\title{Document Object}
\usage{
Document(
  Document.fields = NULL,
  updateTime = NULL,
  fields = NULL,
  name = NULL,
  createTime = NULL
)
}
\arguments{
\item{Document.fields}{The \link{Document.fields} object or list of objects}

\item{updateTime}{Output only}

\item{fields}{The document's fields}

\item{name}{The resource name of the document, for example \code{projects/{project_id}/databases/{database_id}/documents/{document_path}}}

\item{createTime}{Output only}
}
\value{
Document object
}
\description{
Document Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Firestore document. Must not exceed 1 MiB - 4 bytes.
}
\seealso{
Other Document functions: 
\code{\link{Document.fields}()},
\code{\link{projects.databases.documents.createDocument}()},
\code{\link{projects.databases.documents.patch}()}
}
\concept{Document functions}
