% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{GoogleLongrunningOperation}
\alias{GoogleLongrunningOperation}
\title{GoogleLongrunningOperation Object}
\usage{
GoogleLongrunningOperation(
  metadata = NULL,
  response = NULL,
  done = NULL,
  error = NULL,
  name = NULL
)
}
\arguments{
\item{metadata}{Service-specific metadata associated with the operation. It typically contains progress information and common metadata such as create time. Some services might not provide such metadata. Any method that returns a long-running operation should document the metadata type, if any.}

\item{response}{The normal response of the operation in case of success. If the original method returns no data on success, such as \code{Delete}, the response is \code{google.protobuf.Empty}. If the original method is standard \code{Get}/\code{Create}/\code{Update}, the response should be the resource. For other methods, the response should have the type \code{XxxResponse}, where \code{Xxx} is the original method name. For example, if the original method name is \code{TakeSnapshot()}, the inferred response type is \code{TakeSnapshotResponse}.}

\item{done}{If the value is \code{false}, it means the operation is still in progress}

\item{error}{The error result of the operation in case of failure or cancellation}

\item{name}{The server-assigned name, which is only unique within the same service that originally returns it}
}
\value{
GoogleLongrunningOperation object
}
\description{
GoogleLongrunningOperation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This resource represents a long-running operation that is the result of a network API call.
}
\concept{GoogleLongrunningOperation functions}
