% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{Location}
\alias{Location}
\title{Location Object}
\usage{
Location(
  labels = NULL,
  metadata = NULL,
  locationId = NULL,
  name = NULL,
  displayName = NULL
)
}
\arguments{
\item{labels}{Cross-service attributes for the location. For example {'cloud.googleapis.com/region': 'us-east1'}}

\item{metadata}{Service-specific metadata. For example the available capacity at the given location.}

\item{locationId}{The canonical id for this location}

\item{name}{Resource name for the location, which may vary between implementations}

\item{displayName}{The friendly name for this location, typically a nearby city name}
}
\value{
Location object
}
\description{
Location Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A resource that represents Google Cloud Platform location.
}
\concept{Location functions}
