% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{MapValue}
\alias{MapValue}
\title{MapValue Object}
\usage{
MapValue(fields = NULL)
}
\arguments{
\item{fields}{The map's fields. The map keys represent field names. Field names matching the regular expression \verb{__.*__} are reserved. Reserved field names are forbidden except in certain documented contexts. The map keys, represented as UTF-8, must not exceed 1,500 bytes and cannot be empty.}
}
\value{
MapValue object
}
\description{
MapValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map value.
}
\concept{MapValue functions}
