% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore_objects.R
\name{StructuredQuery}
\alias{StructuredQuery}
\title{StructuredQuery Object}
\usage{
StructuredQuery(
  select = NULL,
  from = NULL,
  where = NULL,
  orderBy = NULL,
  startAt = NULL,
  endAt = NULL,
  offset = NULL,
  limit = NULL
)
}
\arguments{
\item{select}{The projection to return}

\item{from}{The collections to query}

\item{where}{The filter to apply}

\item{orderBy}{The order to apply to the query results}

\item{startAt}{A starting point for the query results}

\item{endAt}{A end point for the query results}

\item{offset}{The number of results to skip}

\item{limit}{The maximum number of results to return}
}
\value{
StructuredQuery object
}
\description{
StructuredQuery Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Firestore query.
}
\concept{StructuredQuery functions}
